// webapp/src/app/core/api/v1/model/applicationReportType.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ApplicationReportType {
  applicationTypeId?: number;
  reportType?: number;
  name?: string | null;
  description?: string | null;
  groupType: string;
}